DIGG-67-beta: Updating ContainerPage html structure
