+++ components/Navigation/ContainerNav/index.tsx
@@ -1,129 +1,85 @@
 import Link from "next/link";
-import Heading from "@/components/global/Typography/Heading";
-import { useEffect, useRef, useState } from "react";
+import { useRouter } from "next/router";
+import { useEffect, useState } from "react";
+import { RelatedContainerFragment } from "@/graphql/__generated__/operations";
+import { Button } from "@/components/global/Button";
+import CloseCrossIcon from "@/assets/icons/closeCross.svg";
+import HamburgerIcon from "@/assets/icons/hamburger.svg";
 
 interface ContainerDpDwnProps {
-  menuItems: Anchorlink[];
-  menuHeading: string;
+  related: RelatedContainerFragment[];
+  domain?: DiggDomain;
 }
 
-interface WatchedItem {
-  isActive: boolean;
-  id: string;
-}
+const ContainerNav: React.FC<ContainerDpDwnProps> = ({ related, domain }) => {
+  const [expanded, setExpanded] = useState(false);
 
-const isInView = (element: HTMLElement) => {
-  if (typeof window !== "undefined") {
-    const rect = element.getBoundingClientRect();
-    return rect.top >= 0 && rect.bottom <= window.innerHeight;
-  }
-  return false;
-};
+  useEffect(() => {
+    window.addEventListener("resize", () => setExpanded(false));
+    window.addEventListener("scroll", () => setExpanded(false));
 
-const ContainerNav: React.FC<ContainerDpDwnProps> = ({
-  menuItems,
-  menuHeading,
-}) => {
-  const [latestActiveItem, setLatestActiveItem] = useState<WatchedItem | null>(
-    null,
-  );
-  const [fixedNav, setFixedNav] = useState<{
-    top: boolean;
-    bottom: boolean;
-    bottomPosition: string;
-  }>({
-    top: false,
-    bottom: false,
-    bottomPosition: "",
+    return () => {
+      window.removeEventListener("resize", () => setExpanded(false));
+      window.removeEventListener("scroll", () => setExpanded(false));
+    };
   });
-  let timer = useRef(0);
 
-  const watch = () => {
-    let watchedItems: WatchedItem[] = [];
-    menuItems.map((item) => {
-      const element = document.getElementById(item.id);
+  const { asPath } = useRouter() || {};
 
-      if (element) {
-        watchedItems.push({
-          isActive: isInView(element),
-          id: item.id,
-        });
-      }
-    });
-
-    const latestActiveItem = watchedItems.find((item) => item.isActive) || null;
-
-    if (latestActiveItem != null) {
-      setLatestActiveItem(latestActiveItem);
-    }
+  const isActive = (url: string) => {
+    return asPath === url;
   };
 
-  useEffect(() => {
-    timer.current = window.setInterval(() => {
-      watch();
-    }, 100);
-
-    return () => {
-      window.clearInterval(timer.current);
-    };
-  });
-
-  useEffect(() => {
-    const handleScroll = () => {
-      const containerNav = document.getElementById("containerNav");
-      const navBox = document.getElementById("navBox");
-
-      if (navBox && containerNav) {
-        const navRect = containerNav.getBoundingClientRect();
-        const navBoxRect = navBox.getBoundingClientRect();
-
-        setFixedNav({
-          top: navRect.top < 97 && window.scrollY <= navRect.height,
-          bottom: window.scrollY > navRect.height,
-          bottomPosition: `${navRect.height - navBoxRect.height - 40}px`,
-        });
-      }
-    };
-
-    window.addEventListener("scroll", handleScroll);
-
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, [fixedNav]);
-
   return (
-    <div
-      id="navBox"
-      className={`mb-xl flex flex-col gap-md lg:w-[132px] ${
-        fixedNav.top
-          ? "lg:fixed lg:top-[96px]"
-          : fixedNav.bottom
-          ? "lg:absolute"
-          : "static"
+    <nav
+      className={`xl:col-span-1 xl:col-start-1 xl:row-span-2 xl:flex ${
+        expanded
+          ? `fixed left-none top-none h-screen w-screen overflow-hidden bg-brown-800 px-md py-lg md:relative md:h-fit md:w-fit md:overflow-visible md:bg-transparent md:p-none md:py-none`
+          : "relative"
       }`}
-      style={{ top: fixedNav.bottom ? fixedNav.bottomPosition : "" }}
     >
-      <Heading level={3} size={"xs"} className="text-brown-600">
-        {menuHeading}
-      </Heading>
-      <ul className={`flex w-full flex-col`}>
-        {menuItems.map((item) => (
-          <li key={item.id}>
-            <Link
-              href={`#${item.id}`}
-              className={`inline-flex cursor-pointer p-sm pl-lg text-sm no-underline underline-offset-4 hover:underline ${
-                latestActiveItem?.id === item.id
-                  ? "border-l-[3px] border-pink-600 pl-[18px] font-strong"
-                  : "border-l  border-brown-200 font-normal"
+      <Button
+        iconPosition="left"
+        icon={expanded ? CloseCrossIcon : HamburgerIcon}
+        label={related[0].name}
+        onClick={() => setExpanded(!expanded)}
+        className={`w-full md:w-[320px] xl:hidden`}
+      />
+      <ul
+        className={`flex w-full flex-col bg-white md:absolute md:w-[320px] xl:static xl:flex xl:bg-transparent ${
+          expanded
+            ? "mt-md h-fit max-h-[calc(100vh-92px)] overflow-y-scroll md:mt-none md:max-h-[calc(100vh-228px)]"
+            : "hidden"
+        }`}
+      >
+        {related.map(({ name, slug }) => {
+          const url = `${domain ? "/" + domain : ""}${slug}`;
+          return (
+            <li
+              className={`${
+                isActive(url) ? " bg-brown-900 text-white" : "text-brown-600"
               }`}
+              key={slug}
             >
-              {item.text}
-            </Link>
-          </li>
-        ))}
+              <Link
+                href={url}
+                className={`inline-flex w-full px-md py-sm no-underline ${
+                  isActive(url) ? "cursor-default" : "hover:underline"
+                }`}
+                aria-disabled={isActive(url)}
+                onClick={() => {
+                  setExpanded(false);
+                  window.scrollTo({ top: 0, behavior: "smooth" });
+                }}
+                scroll={false}
+              >
+                {name}
+              </Link>
+            </li>
+          );
+        })}
       </ul>
-    </div>
+    </nav>
   );
 };
